+++ ViewModel/Search/ExtraData.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Stape\Gtm\ViewModel\Search;
+
+use Magento\Framework\View\Element\Block\ArgumentInterface;
+use Stape\Gtm\ViewModel\Category\ExtraData as CategoryExtraData;
+use Stape\Gtm\ViewModel\DatalayerInterface;
+
+class ExtraData extends CategoryExtraData implements ArgumentInterface, DatalayerInterface
+{
+
+}